Reduce the complexity of html report generation
For the exsiting way to generate a html report, letp
depends on the generated json file as well as the
junit xml file. The framework parses some configs
from the json file and the test results from the
junit xml file. This is redudant because the test
results are also included in the json file

Parsing the test results from the json file is
already supported becase of LETP-468

Resolves: LETP-402
Change-Id: Ib83ced39129f73e79d7e517c68e473e9e3bdfe1a