Correct configLeTP.sh directory logic
If the containing directory ends in -qa, configLeTP will get
confused and set QA_ROOT to the parent directory instead of the QA
directory.  Adjust the logic which determines the QA_ROOT to correct
for this.

The letp directory will usually be alongside the legato directory.
Detect if this is the case, and if so, set LEGATO_ROOT to the
correct directory.

Resolves: LETP-515
Change-Id: I59d6b793363f834cef8acf0b5efaae3c0f734eb1