Create LeTP container environment and wrapper
Introduce a Docker container for a standardized LeTP runtime
environment, and create a wrapper script to manage the configuration of
the container and execution of tests.  See container/README.md for
details.

Also enhance the LeTP configuration XML include behaviour to allow
resolution of environment variables in the file paths, and to allow the
default top-level testbench.xml file to be set via the LETP_CONFIG_XML
environment variable.

Finally, change the serial port handling behaviour of LeTP a bit, as
some of the assumptions it makes about serial devices don't hold inside
the container environment.

Resolves: LETP-523
Change-Id: If8d34b3c0e54b54ae76b66c3290a610b89a281b8